backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
openai==1.61.0
python-dotenv==1.0.0
httpx==0.27.0
